rename extensions
