+++ .eslintrc.json
@@ -925,14 +925,6 @@
 				]
 			},
 			{
-				"target": "**/extensions/gitpod-web/**",
-				"restrictions": "!@gitpod/supervisor-api-grpc/**"
-			},
-			{
-				"target": "**/extensions/gitpod-remote-ssh/**",
-				"restrictions": "!@gitpod/supervisor-api-grpc/**"
-			},
-			{
 				"target": "**/extensions/**",
 				"restrictions": "**/*"
 			},
