+++ .gitpod.yml
@@ -17,7 +17,8 @@ tasks:
       export NODE_ENV=development
       export VSCODE_DEV=1
       gp sync-await init
-      yarn gitpod:watch
+      cd ./extensions
+      yarn watch
     name: watch extension
   - command: |
       export NODE_ENV=development
