--- extensions/gitpod-desktop/README.md
@@ -1,5 +1,0 @@
-# Gitpod Desktop integration
-
-## Features
-
-This extension provides support for Gitpod Desktop integration.
