--- extensions/gitpod-desktop/package.json
@@ -1,47 +1,0 @@
-{
-	"name": "gitpod-desktop",
-	"displayName": "%displayName%",
-	"description": "%description%",
-	"publisher": "gitpod",
-	"version": "0.0.14",
-	"license": "MIT",
-	"preview": true,
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/gitpod-io/vscode.git"
-	},
-	"engines": {
-		"vscode": "^1.58.2"
-	},
-	"categories": [
-		"Other"
-	],
-	"extensionKind": [
-		"ui"
-	],
-	"activationEvents": [
-		"*",
-		"onCommand:gitpod-desktop.api.autoTunnel"
-	],
-	"main": "./out/extension.js",
-	"scripts": {
-		"compile": "gulp compile-extension:gitpod-desktop",
-		"watch": "gulp watch-extension:gitpod-desktop",
-		"vscode:prepublish": "npm run compile"
-	},
-	"devDependencies": {
-		"@types/node": "^10.12.21",
-		"@types/node-fetch": "^2.5.12",
-		"@types/tmp": "^0.2.1"
-	},
-	"dependencies": {
-		"@gitpod/local-app-api-grpcweb": "ak-vscode-desktop",
-		"@improbable-eng/grpc-web-node-http-transport": "^0.14.0",
-		"node-fetch": "^2.6.1",
-		"tmp": "^0.2.1",
-		"vscode-nls": "^5.0.0"
-	},
-	"extensionDependencies": [
-		"ms-vscode-remote.remote-ssh"
-	]
-}
