--- extensions/gitpod-desktop/package.nls.json
@@ -1,4 +1,0 @@
-{
-	"displayName": "Gitpod Desktop",
-	"description": "Gitpod Desktop Support"
-}
