--- extensions/gitpod-remote-ssh/.vscodeignore
@@ -1,9 +1,0 @@
-.gitignore
-src/**
-!src/common/config.json
-out/**
-build/**
-extension.webpack.config.js
-extension-browser.webpack.config.js
-tsconfig.json
-yarn.lock
