--- extensions/gitpod-remote-ssh/package.json
@@ -1,41 +1,0 @@
-{
-	"name": "gitpod-remote-ssh",
-	"displayName": "%displayName%",
-	"description": "%description%",
-	"publisher": "gitpod",
-	"version": "0.0.14",
-	"license": "MIT",
-	"preview": true,
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/gitpod-io/vscode.git"
-	},
-	"engines": {
-		"vscode": "^1.58.2"
-	},
-	"categories": [
-		"Other"
-	],
-	"extensionKind": [
-		"workspace"
-	],
-	"activationEvents": [
-		"*"
-	],
-	"main": "./out/extension.js",
-	"contributes": {},
-	"scripts": {
-		"compile": "gulp compile-extension:gitpod-remote-ssh",
-		"watch": "gulp watch-extension:gitpod-remote-ssh",
-		"vscode:prepublish": "npm run compile"
-	},
-	"devDependencies": {
-		"@types/node": "^10.12.21"
-	},
-	"dependencies": {
-		"vscode-nls": "^5.0.0"
-	},
-	"extensionDependencies": [
-		"gitpod.gitpod"
-	]
-}
