--- extensions/gitpod-remote-ssh/package.nls.json
@@ -1,4 +1,0 @@
-{
-	"displayName": "Gitpod Remote SSH",
-	"description": "Gitpod Remote SSH Support"
-}
