+++ extensions/gitpod-remote/.vscodeignore
