+++ extensions/gitpod-remote/package.json
@@ -1,14 +1,18 @@
 {
-  "name": "gitpod-web",
+  "name": "gitpod-remote-ssh",
   "displayName": "%displayName%",
   "description": "%description%",
   "publisher": "gitpod",
-  "version": "0.0.1",
+  "version": "0.0.16",
   "license": "MIT",
+  "preview": true,
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/gitpod-io/vscode.git"
+  },
   "engines": {
     "vscode": "^1.58.2"
   },
-  "enableProposedApi": true,
   "categories": [
     "Other"
   ],
@@ -16,9 +20,7 @@
     "workspace"
   ],
   "activationEvents": [
-    "*",
-    "onAuthenticationRequest:gitpod",
-    "onAuthenticationRequest:github"
+    "*"
   ],
   "main": "./out/extension.js",
   "contributes": {
@@ -112,209 +114,9 @@
         "command": "gitpod.openInBrowser",
         "title": "%openInBrowser%",
         "enablement": "gitpod.inWorkspace == true && gitpod.UIKind == 'desktop'"
-      },
-      {
-        "command": "gitpod.ports.openBrowser",
-        "title": "%openBrowser%",
-        "icon": "$(globe)"
-      },
-      {
-        "command": "gitpod.ports.retryAutoExpose",
-        "title": "%retryAutoExpose%",
-        "icon": "$(refresh)"
-      },
-      {
-        "command": "gitpod.ports.preview",
-        "title": "%openPreview%",
-        "icon": "$(open-preview)"
-      },
-      {
-        "command": "gitpod.ports.makePrivate",
-        "title": "%makePrivate%",
-        "icon": "$(lock)"
-      },
-      {
-        "command": "gitpod.ports.makePublic",
-        "title": "%makePublic%",
-        "icon": "$(unlock)"
-      },
-      {
-        "command": "gitpod.ports.tunnelNetwork",
-        "title": "%tunnelNetwork%",
-        "icon": "$(eye-closed)"
-      },
-      {
-        "command": "gitpod.ports.tunnelHost",
-        "title": "%tunnelHost%",
-        "icon": "$(eye)"
-      },
-      {
-        "command": "gitpod.extensions.addToConfig",
-        "title": "%addToConfig%"
-      },
-      {
-        "command": "gitpod.dev.connectLocalApp",
-        "title": "%connectLocalApp%"
-      },
-      {
-        "command": "gitpod.dev.enableForwardedPortsView",
-        "title": "%enableForwardedPortsView%"
       }
     ],
     "menus": {
-      "menuBar/home": [
-        {
-          "command": "gitpod.open.context",
-          "group": "gitpod@10"
-        },
-        {
-          "command": "gitpod.open.documentation",
-          "group": "gitpod@20"
-        },
-        {
-          "command": "gitpod.takeSnapshot",
-          "group": "gitpod@30",
-          "when": "gitpod.workspaceOwned == true"
-        },
-        {
-          "command": "gitpod.shareWorkspace",
-          "group": "gitpod@40",
-          "when": "gitpod.workspaceOwned == true && gitpod.workspaceShared == false"
-        },
-        {
-          "command": "gitpod.stopSharingWorkspace",
-          "group": "gitpod@40",
-          "when": "gitpod.workspaceOwned == true && gitpod.workspaceShared == true"
-        },
-        {
-          "command": "gitpod.stop.ws",
-          "group": "gitpod@50",
-          "when": "gitpod.workspaceOwned == true"
-        },
-        {
-          "command": "gitpod.openInStable",
-          "group": "gitpod@60"
-        },
-        {
-          "command": "gitpod.openInInsiders",
-          "group": "gitpod@70"
-        }
-      ],
-      "accounts/context": [
-        {
-          "command": "gitpod.open.settings",
-          "group": "navigation@10"
-        },
-        {
-          "command": "gitpod.open.accessControl",
-          "group": "navigation@20"
-        },
-        {
-          "command": "gitpod.upgradeSubscription",
-          "group": "navigation@30",
-          "when": "gitpod.workspaceOwned == true"
-        },
-        {
-          "command": "gitpod.ExtendTimeout",
-          "group": "navigation@50",
-          "when": "gitpod.workspaceOwned == true"
-        }
-      ],
-      "menuBar/help": [
-        {
-          "command": "gitpod.open.twitter",
-          "group": "z_about2@10"
-        },
-        {
-          "command": "gitpod.open.discord",
-          "group": "z_about2@20"
-        },
-        {
-          "command": "gitpod.open.discourse",
-          "group": "z_about2@30"
-        },
-        {
-          "command": "gitpod.reportIssue",
-          "group": "z_about2@40"
-        }
-      ],
-      "extension/context": [
-        {
-          "command": "gitpod.extensions.addToConfig",
-          "group": "2_configure"
-        }
-      ],
-      "view/item/context": [
-        {
-          "command": "gitpod.ports.tunnelNetwork",
-          "when": "view == gitpod.workspace && viewItem =~ /host/ && viewItem =~ /tunneled/",
-          "group": "inline@1"
-        },
-        {
-          "command": "gitpod.ports.tunnelHost",
-          "when": "view == gitpod.workspace && viewItem =~ /network/ && viewItem =~ /tunneled/",
-          "group": "inline@1"
-        },
-        {
-          "command": "gitpod.ports.makePublic",
-          "when": "view == gitpod.workspace && viewItem =~ /private/",
-          "group": "inline@2"
-        },
-        {
-          "command": "gitpod.ports.makePrivate",
-          "when": "view == gitpod.workspace && viewItem =~ /public/",
-          "group": "inline@2"
-        },
-        {
-          "command": "gitpod.ports.preview",
-          "when": "view == gitpod.workspace && viewItem =~ /exposed/ || viewItem =~ /tunneled/",
-          "group": "inline@3"
-        },
-        {
-          "command": "gitpod.ports.openBrowser",
-          "when": "view == gitpod.workspace && viewItem =~ /exposed/ || viewItem =~ /tunneled/",
-          "group": "inline@4"
-        },
-        {
-          "command": "gitpod.ports.retryAutoExpose",
-          "when": "view == gitpod.workspace && viewItem =~ /failed/",
-          "group": "inline@5"
-        }
-      ],
-      "commandPalette": [
-        {
-          "command": "gitpod.ports.preview",
-          "when": "false"
-        },
-        {
-          "command": "gitpod.ports.openBrowser",
-          "when": "false"
-        },
-        {
-          "command": "gitpod.ports.retryAutoExpose",
-          "when": "false"
-        },
-        {
-          "command": "gitpod.ports.makePublic",
-          "when": "false"
-        },
-        {
-          "command": "gitpod.ports.makePrivate",
-          "when": "false"
-        },
-        {
-          "command": "gitpod.ports.tunnelNetwork",
-          "when": "false"
-        },
-        {
-          "command": "gitpod.ports.tunnelHost",
-          "when": "false"
-        },
-        {
-          "command": "gitpod.dev.connectLocalApp",
-          "when": "gitpod.localAppConnected == false"
-        }
-      ],
       "statusBar/remoteIndicator": [
         {
           "command": "gitpod.stop.ws",
@@ -407,38 +209,16 @@
           "when": "gitpod.inWorkspace == true && gitpod.UIKind == 'desktop'"
         }
       ]
-    },
-    "views": {
-      "remote": [
-        {
-          "id": "gitpod.workspace",
-          "name": "Gitpod Workspace"
-        }
-      ]
-    },
-    "authentication": [
-      {
-        "id": "gitpod",
-        "label": "Gitpod"
-      },
-      {
-        "id": "github",
-        "label": "GitHub"
-      }
-    ]
+    }
   },
   "scripts": {
     "vscode:prepublish": "yarn run compile --cwd .."
   },
   "devDependencies": {
-    "@types/node": "^10.12.21",
-    "@types/node-fetch": "^2.5.8"
+    "@types/node": "^10.12.21"
   },
   "dependencies": {
     "gitpod-shared": "0.0.1",
-    "node-fetch": "^2.6.1",
-    "uuid": "^8.3.1",
-    "vscode-jsonrpc": "^5.0.1",
     "vscode-nls": "^5.0.0"
   }
 }
