+++ extensions/gitpod-remote/package.nls.json
@@ -1,16 +1,6 @@
 {
-	"displayName": "Gitpod Web",
-	"description": "Gitpod Web Support",
-	"openBrowser": "Open Browser",
-	"openPreview": "Open Preview",
-	"makePrivate": "Make Private",
-	"makePublic": "Make Public",
-	"tunnelNetwork": "Tunnel on all interfaces",
-	"tunnelHost": "Tunnel on localhost",
-	"addToConfig": "Add to .gitpod.yml",
-	"connectLocalApp": "Developer: Connect to Local Gitpod App",
-	"enableForwardedPortsView": "Developer: Enable Tunneled Ports View",
-	"retryAutoExpose": "Retry to expose",
+	"displayName": "Gitpod Remote",
+	"description": "Gitpod Remote Support",
 	"openDashboard": "Gitpod: Open Dashboard",
 	"openAccessControl": "Gitpod: Open Access Control",
 	"openSettings": "Gitpod: Open Settings",
