+++ extensions/gitpod-shared/tsconfig.json
@@ -3,6 +3,7 @@
 	"compilerOptions": {
 		"outDir": "./out",
 		"rootDir": "./src",
+		"composite": true,
 		"esModuleInterop": true
 	}
 }
