+++ extensions/gitpod/extension.webpack.config.js
@@ -7,7 +7,6 @@
 
 'use strict';
 
-const path = require('path');
 const withDefaults = require('../shared.webpack.config');
 
 module.exports = withDefaults({
