+++ extensions/package.json
@@ -1,4 +1,5 @@
 {
+  "private": true,
   "name": "vscode-extensions",
   "version": "0.0.1",
   "license": "MIT",
@@ -7,10 +8,20 @@
     "typescript": "4.4.3"
   },
   "scripts": {
-    "postinstall": "node ./postinstall"
+    "postinstall": "node ./postinstall",
+    "compile": "../node_modules/.bin/tsc -b",
+    "watch": "../node_modules/.bin/tsc -b -w"
   },
   "devDependencies": {
     "esbuild": "^0.11.12",
     "vscode-grammar-updater": "^1.0.3"
+  },
+  "workspaces": {
+    "packages": [
+      "gitpod-shared",
+      "gitpod-remote",
+      "gitpod-web",
+      "gitpod"
+    ]
   }
 }
